requirements.txt
asgiref==3.7.2
Django==4.2.5
django-cors-headers==4.2.0
numpy==1.25.2
opencv-python==4.8.0.76
Pillow==10.0.0
sqlparse==0.4.4
